+++ lua/plugins/noice.lua
@@ -50,27 +50,6 @@ return {
       end,
       desc = "Noice Errors",
     },
-    -- TODO: conflict with coc
-    -- {
-    --   "<C-d>",
-    --   function()
-    --     if not require("noice.lsp").scroll(4) then return "<C-d>" end
-    --   end,
-    --   silent = true,
-    --   expr = true,
-    --   desc = "Scroll forward",
-    --   mode = { "i", "n", "s" }
-    -- },
-    -- {
-    --   "<C-u>",
-    --   function()
-    --     if not require("noice.lsp").scroll(-4) then return "<C-u>" end
-    --   end,
-    --   silent = true,
-    --   expr = true,
-    --   desc = "Scroll backward",
-    --   mode = { "i", "n", "s" }
-    -- },
   },
 }